pony/src/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
certifi==2024.8.30
charset-normalizer==3.4.0
complex==0.2.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.1
gensim==4.3.3
grad-cam==1.5.4
idna==3.10
imageio==2.36.1
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
matplotlib==3.9.3
networkx==3.2.1
numba==0.60.0
numpy==1.23.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
six==1.17.0
smart-open==7.0.5
tensorboardX==2.6.2.2
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.8.30
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchmetrics==0.6.0
torchvision==0.16.2+cu118
tqdm==4.67.1
ttach==0.0.3
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.17.0
zipp==3.21.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
ftfy==6.2.0
mkl-service==2.4.0
regex==2024.5.15
wcwidth==0.2.13
openpyxl==3.1.5
pandas==2.0.3
